requirements.txt
yara-python
psutil